Lab2: some renaming
