+++ apps/src/code-studio/redux.js
@@ -18,7 +18,7 @@ import currentUser from '@cdo/apps/templates/currentUserRedux';
 import arrowDisplay from '@cdo/apps/templates/arrowDisplayRedux';
 import teacherPanel from '@cdo/apps/code-studio/teacherPanelRedux';
 import microBit from '../lib/kits/maker/microBitRedux';
-import lab from '../labs/labRedux';
+import lab from '../lab2/lab2Redux';
 
 registerReducers({
   header,
