+++ apps/src/lab2/ErrorBoundary.tsx
