+++ apps/src/lab2/constants.ts
