+++ apps/src/lab2/lab2Registry.ts
