+++ apps/src/lab2/progress/ConditionsChecker.ts
