+++ apps/src/lab2/projects/utils.ts
