+++ apps/src/lab2/types.ts
