+++ apps/src/lab2/views/Lab2Wrapper.module.scss
