+++ apps/src/sites/studio/pages/levels-lab2-main.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Lab2 from '@cdo/apps/labs/views/lab2';
+import Lab2 from '@cdo/apps/lab2/views/Lab2';
 
 $(document).ready(function () {
   ReactDOM.render(<Lab2 />, document.getElementById('lab2-container'));
