+++ apps/src/sites/studio/pages/musiclab/index.js
@@ -4,10 +4,10 @@ import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {getStore} from '@cdo/apps/redux';
 import MusicLabView from '@cdo/apps/music/views/MusicView';
-import ProjectContainer from '@cdo/apps/labs/projects/ProjectContainer';
-import ErrorBoundary from '@cdo/apps/labs/ErrorBoundary';
+import ProjectContainer from '@cdo/apps/lab2/projects/ProjectContainer';
+import ErrorBoundary from '@cdo/apps/lab2/ErrorBoundary';
 import {logError} from '@cdo/apps/music/utils/MusicMetrics';
-import {ErrorFallbackPage} from '@cdo/apps/labs/views/Lab2Wrapper';
+import {ErrorFallbackPage} from '@cdo/apps/lab2/views/Lab2Wrapper';
 
 $(document).ready(function () {
   const channelId = document.querySelector('script[data-channelid]').dataset
