+++ apps/src/standaloneVideo/StandaloneVideo.tsx
@@ -13,8 +13,8 @@ import {
   sendSuccessReport,
   navigateToNextLevel,
 } from '@cdo/apps/code-studio/progressRedux';
-import {LabState} from '@cdo/apps/labs/labRedux';
-import {VideoLevelData} from '@cdo/apps/labs/types';
+import {LabState} from '@cdo/apps/lab2/lab2Redux';
+import {VideoLevelData} from '@cdo/apps/lab2/types';
 import standaloneVideoLocale from './locale';
 import styles from './video.module.scss';
